refactor(live-painting): split live painting file
- separates various elements
- provides reusable components for various elements
